requirements.txt
# Google Cloud dependencies
google-cloud-aiplatform>=1.38.0
google-cloud-storage>=2.10.0
google-auth>=2.23.0

# Web scraping
beautifulsoup4>=4.12.0
requests>=2.31.0
lxml>=4.9.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# Configuration and utilities
pyyaml>=6.0
python-dotenv>=1.0.0
tqdm>=4.66.0

# PDF generation
reportlab>=4.0.0
pillow>=10.0.0

# Web framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
jinja2>=3.1.2

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Type hints and code quality
mypy>=1.5.0
black>=23.0.0
flake8>=6.1.0

# Additional utilities
python-dateutil>=2.8.0
jsonschema>=4.19.0
